상세페이지에서 추천수 노출

+++ src/main/webapp/WEB-INF/jsp/usr/article/modify.jsp
@@ -1,5 +1,4 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 <c:set var="pageTitle" value="게시물 수정" />
 <%@ include file="../common/head.jspf"%>
@@ -8,7 +7,7 @@
 	<div class="container mx-auto px-3">
 		<div class="table-box-type-1">
 			<form class="table-box-type-1" method="POST" action="../article/doModify">
-				<input type="hidden" name="id" value="${article.id}"/>
+				<input type="hidden" name="id" value="${article.id}" />
 				<table>
 					<colgroup>
 						<col width="200" />
@@ -17,7 +16,7 @@
 						<tr>
 							<th>번호</th>
 							<td>
-							<div class="badge badge-primary"> ${article.id}</div>
+								<div class="badge badge-primary">${article.id}</div>
 							</td>
 						</tr>
 						<tr>
@@ -29,38 +28,43 @@
 							<td>${article.forPrintType1UpdateDate}</td>
 						</tr>
 						<tr>
+							<th>조회</th>
+							<td><span class="text-blue-700">${article.hitCount}</span></td>
+						</tr>
+						<tr>
+							<th>추천수</th>
+							<td>
+								<span class="text-blue-700">${article.extra__goodReactionPoint}</span>
+							</td>
+						</tr>
+						<tr>
 							<th>작성자</th>
 							<td>${article.extra__writerName}</td>
 						</tr>
 						<tr>
 							<th>제목</th>
-							<td><textarea class="w-full text-black" name="body" rows="1">${article.body}</textarea>
-							</td>
+							<td><textarea class="w-full text-black" name="body" rows="1">${article.body}</textarea></td>
 						</tr>
 						<tr>
 							<th>내용</th>
-							<td><textarea class="w-full text-black" name="body"
-									rows="10">${article.body}</textarea></td>
+							<td><textarea class="w-full text-black" name="body" rows="10">${article.body}</textarea></td>
 						</tr>
 						<tr>
 							<th>수정</th>
 							<td><input type="submit" class="btn btn-primary" value="수정" />
-								<button type="button"class="btn btn-outline btn-success"   onclick="history.back();">뒤로가기</button></td>
+								<button type="button" class="btn btn-outline btn-success" onclick="history.back();">뒤로가기</button></td>
 						</tr>
 					</tbody>
 				</table>
 		</div>
 		<div class="button">
 			<button type="button" onclick="history.back();">뒤로가기</button>
-			<a class="btn-text-link" href="../article/Modify?id=${article.id}">게시물
-				수정</a>
+			<a class="btn-text-link" href="../article/Modify?id=${article.id}">게시물 수정</a>
 			<c:if test="${article.extra__actorCanDelete}">
-				<a class="btn-text-link"
-					onclick="if (confirm('정말 삭제하시겠습니까?') == false) return false;"
-					href="../article/delete/delete?id=${article.id}"> 게시물 삭제</a>
+				<a class="btn-text-link" onclick="if (confirm('정말 삭제하시겠습니까?') == false) return false;" href="../article/delete/delete?id=${article.id}"> 게시물 삭제</a>
 			</c:if>
 		</div>
-		
+
 	</div>
 </section>
